requirements.txt
amqp==5.1.1
apimatic-core==0.2.4
apimatic-core-interfaces==0.1.4
apimatic-requests-client-adapter==0.1.5
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
black==23.7.0
boto3==1.26.59
botocore==1.29.165
CacheControl==0.12.14
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
cryptography==41.0.3
deprecation==2.1.0
Django==4.2.3
django-admin-autocomplete-filter==0.7.1
django-admin-inline-paginator==0.4.0
django-admin-rangefilter==0.9.0
django-appconf==1.0.5
django-celery-beat==2.5.0
django-celery-results==2.0.0
django-ckeditor==6.7.0
django-cryptography==1.1
django-db-logger==0.1.12
django-js-asset==2.1.0
django-multiupload==0.6.1
django-redis==5.4.0
django-storages==1.13.2
django-timezone-field==6.0
enum34==1.1.10
et-xmlfile==1.1.0
Faker==19.2.0
idna==3.4
jmespath==1.0.1
jsonpickle==3.0.2
jsonpointer==2.4
kombu==5.3.1
msgpack==1.0.5
mypy-extensions==1.0.0
openpyxl==3.1.2
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycparser==2.21
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==4.6.0
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
squareup==29.0.0.20230720
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
